Moving IntoReveal impls to avoid pubs in node types
